Bumping Chart versions as pacman k8s version is much newer than our c… (#20)
…harts

- Bumps chart version as pacman k8s version is much newer than our
charts

### PR Description
Please replace this with a high level description of what this PR is
supposed to accomplish.

### Checklist

- [ ] Have you reviewed and updated the chart default values if
necessary?
- [ ] Have you reviewed and updated the chart documentation if
necessary?
- [ ] Does your branch follow the naming convention of
`{chartNameWithDashes}-v{versionString}-{optionalPatchVersion}`?
 - [ ] Have you bumped the version in the chart's `Chart.yaml`?

### Tagged Releases
Please remember to make a tagged release after merging your PR that:

 - Has a tag name that matches your PR branch name (see above)
 - Has a description that summarizes the changes made

This makes it possible to use previous versions of the charts maintained
here as new releases are published. Please see the release history of
this repository for examples.