Keess v0.1.0 (#23)
### PR Description
This PR creates a new chart for the new secrets and configmaps
synchronizer named keess.

### Checklist

- [X] Have you reviewed and updated the chart default values if
necessary?
- [X] Have you reviewed and updated the chart documentation if
necessary?
- [X] Does your branch follow the naming convention of
`{chartNameWithDashes}-v{versionString}-{optionalPatchVersion}`?
 - [X] Have you bumped the version in the chart's `Chart.yaml`?

### Tagged Releases
Please remember to make a tagged release after merging your PR that:

 - Has a tag name that matches your PR branch name (see above)
 - Has a description that summarizes the changes made

This makes it possible to use previous versions of the charts maintained
here as new releases are published. Please see the release history of
this repository for examples.