Feat/legacy elastic stack (#25)
### PR Description
Copy the chart from https://github.com/helm/charts since it is a
deprecated chart and update the apiVersion. I separated the changes into
different commits to make it easier to review this. The first copy is
just the copy and the second is the actual change.

### Checklist

- [ ] Have you reviewed and updated the chart default values if
necessary?
- [ ] Have you reviewed and updated the chart documentation if
necessary?
- [ ] Does your branch follow the naming convention of
`{chartNameWithDashes}-v{versionString}-{optionalPatchVersion}`?
   - No, but I'll create a release after it is merged
 - [ ] Have you bumped the version in the chart's `Chart.yaml`?

### Tagged Releases
Please remember to make a tagged release after merging your PR that:

 - Has a tag name that matches your PR branch name (see above)
 - Has a description that summarizes the changes made

This makes it possible to use previous versions of the charts maintained
here as new releases are published. Please see the release history of
this repository for examples.

---------

Signed-off-by: Igor Valente Blackman <igor.blackman@gmail.com>