fix(clusterapi-resources): Add checksum for kubeadmconfig to trigger MachineDeployment rollout (#36)
### PR Description
If we update the content of `kubeadmConfigSpec` it will trigger rollout
of the Control planes but it won't for MachineDeployments since there
wasn't a change on the resource itself. This PR adds an annotation with
the checksum of kubeadmconfig to force a rollout of the
MachineDeployment in case there is a change in the kubeadmconfig.

### Checklist

- [ ] Have you reviewed and updated the chart default values if
necessary?
- [ ] Have you reviewed and updated the chart documentation if
necessary?
- [ ] Does your branch follow the naming convention of
`{chartNameWithDashes}-v{versionString}-{optionalPatchVersion}`?
 - [x] Have you bumped the version in the chart's `Chart.yaml`?

### Tagged Releases
Please remember to make a tagged release after merging your PR that:

 - Has a tag name that matches your PR branch name (see above)
 - Has a description that summarizes the changes made

This makes it possible to use previous versions of the charts maintained
here as new releases are published. Please see the release history of
this repository for examples.

Signed-off-by: Igor Valente Blackman <igor.blackman@gmail.com>