feat(clusterapi-resources): Add multipath.conf and reverts PR#36 (#37)
### PR Description
Adds `multipath.conf` and reverts the last change that didn't have the
desired effect, turning it unnecessary.

### Checklist

- [ ] Have you reviewed and updated the chart default values if
necessary?
- [ ] Have you reviewed and updated the chart documentation if
necessary?
- [ ] Does your branch follow the naming convention of
`{chartNameWithDashes}-v{versionString}-{optionalPatchVersion}`?
 - [ ] Have you bumped the version in the chart's `Chart.yaml`?

### Tagged Releases
Please remember to make a tagged release after merging your PR that:

 - Has a tag name that matches your PR branch name (see above)
 - Has a description that summarizes the changes made

This makes it possible to use previous versions of the charts maintained
here as new releases are published. Please see the release history of
this repository for examples.

---------

Signed-off-by: Igor Valente Blackman <igor.blackman@gmail.com>