fix(clusterapi-resources): Reload multipathd if multipath is enabled (#41)
### PR Description
Our chart allows the config of multipath, but since the file is created
after the service is started, it never loads the new config. This PR
address this issue by reloading multipathd as a pre kubeadm command.

### Checklist

- [x] Have you reviewed and updated the chart default values if
necessary?
- [x] Have you reviewed and updated the chart documentation if
necessary?
- [ ] Does your branch follow the naming convention of
`{chartNameWithDashes}-v{versionString}-{optionalPatchVersion}`?
 - [x] Have you bumped the version in the chart's `Chart.yaml`?

### Tagged Releases
Please remember to make a tagged release after merging your PR that:

 - Has a tag name that matches your PR branch name (see above)
 - Has a description that summarizes the changes made

This makes it possible to use previous versions of the charts maintained
here as new releases are published. Please see the release history of
this repository for examples.

---------

Signed-off-by: Igor Valente Blackman <igor.blackman@gmail.com>