Update VSphere CPI to 1.32.2 (#48)
### PR Description
In preparation for the upgrade of the Kube clusters, we had to upgrade
our MC providers.

Updates the chart version and appVersion, since the wasn't any change
needed to use CAPI 1.9.6 and the CAPV 1.12.0 only updated the VSphere
CPI version to 1.32.2.

### Checklist

- [x] Have you reviewed and updated the chart default values if
necessary?
- [ ] Have you reviewed and updated the chart documentation if
necessary?
- [ ] Does your branch follow the naming convention of
`{chartNameWithDashes}-v{versionString}-{optionalPatchVersion}`?
 - [x] Have you bumped the version in the chart's `Chart.yaml`?

### Tagged Releases
Please remember to make a tagged release after merging your PR that:

 - Has a tag name that matches your PR branch name (see above)
 - Has a description that summarizes the changes made

This makes it possible to use previous versions of the charts maintained
here as new releases are published. Please see the release history of
this repository for examples.