Move VSphere resourceSet to its own with reconcile (#49)
### PR Description
In order to not have to manage the VSphere resources deployed via
`ClusterResourceSet` outside the pacman repo, this PR moves them to
their own `ClusterResourceSet` with `strategy: Reconcile`.

I decided to bump the major version because it changes an existing
resource and the new one will be reconciling which it didn't do before.

### Checklist

- [x] Have you reviewed and updated the chart default values if
necessary?
- [ ] Have you reviewed and updated the chart documentation if
necessary?
- [ ] Does your branch follow the naming convention of
`{chartNameWithDashes}-v{versionString}-{optionalPatchVersion}`?
 - [x] Have you bumped the version in the chart's `Chart.yaml`?

### Tagged Releases
Please remember to make a tagged release after merging your PR that:

 - Has a tag name that matches your PR branch name (see above)
 - Has a description that summarizes the changes made

This makes it possible to use previous versions of the charts maintained
here as new releases are published. Please see the release history of
this repository for examples.

---------

Signed-off-by: Igor Valente Blackman <igor.blackman@gmail.com>