Make additionalDisksGiB default to be unset.
VsphereMachineTemplates are immutable objects. So even adding this attribute set to an empty list triggers an error on Kubernetes, because it is considered a change in the spec.

This will trigger errors for existing VsphereMachineTemplates that don't need extra disks, and have no change at all.

So instead of adding that empty default, we omit the attribute when unset, and document its existence.