Add support for additional disks in VSphereMachineTemplate (#53)
Add support for additional disks in VSphereMachineTemplate, so we can
add extra disks to nodes created by Cluster API on VSphere.

It does that by supporting the additionalDisksGiB attribute, but
omitting it in the rendered object if value is unset.

This is because VsphereMachineTemplates are immutable objects. So adding
this attribute set to an empty list on an existent object triggers an
error on Kubernetes, because it is considered a change in the spec.

This will trigger errors for existing VsphereMachineTemplates that don't
need extra disks, and have no change at all, causing us trouble. So
instead of adding that empty default, we omit the attribute when unset,
and document its existence.